+++ src/main/java/shadows/apotheosis/adventure/affix/effect/EnlightenedAffix.java
@@ -3,25 +3,32 @@ package shadows.apotheosis.adventure.affix.effect;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import com.google.gson.JsonObject;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import net.minecraft.ChatFormatting;
-import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
-import net.minecraft.util.GsonHelper;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
 import shadows.apotheosis.adventure.AdventureConfig;
 import shadows.apotheosis.adventure.affix.Affix;
-import shadows.apotheosis.adventure.affix.AffixHelper;
 import shadows.apotheosis.adventure.affix.AffixType;
+import shadows.apotheosis.adventure.affix.socket.gem.bonus.GemBonus;
 import shadows.apotheosis.adventure.loot.LootCategory;
 import shadows.apotheosis.adventure.loot.LootRarity;
 import shadows.placebo.util.StepFunction;
 
 public class EnlightenedAffix extends Affix {
+
+	//Formatter::off
+	public static final Codec<EnlightenedAffix> CODEC = RecordCodecBuilder.create(inst -> inst
+		.group(
+			GemBonus.VALUES_CODEC.fieldOf("values").forGetter(a -> a.values))
+			.apply(inst, EnlightenedAffix::new)
+		);
+	//Formatter::on
 
 	protected final Map<LootRarity, StepFunction> values;
 
@@ -49,24 +56,6 @@ public class EnlightenedAffix extends Affix {
 			return InteractionResult.SUCCESS;
 		}
 		return super.onItemUse(stack, rarity, level, ctx);
-	}
-
-	public static Affix read(JsonObject obj) {
-		var values = AffixHelper.readValues(GsonHelper.getAsJsonObject(obj, "values"));
-		return new EnlightenedAffix(values);
-	}
-
-	public JsonObject write() {
-		return new JsonObject();
-	}
-
-	public void write(FriendlyByteBuf buf) {
-		buf.writeMap(this.values, (b, key) -> b.writeUtf(key.id()), (b, func) -> func.write(b));
-	}
-
-	public static Affix read(FriendlyByteBuf buf) {
-		Map<LootRarity, StepFunction> values = buf.readMap(b -> LootRarity.byId(b.readUtf()), b -> StepFunction.read(b));
-		return new EnlightenedAffix(values);
 	}
 
 }