+++ src/main/java/shadows/apotheosis/adventure/affix/effect/ExecutingAffix.java
@@ -3,13 +3,12 @@ package shadows.apotheosis.adventure.affix.effect;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import com.google.gson.JsonObject;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import net.minecraft.ChatFormatting;
-import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.util.GsonHelper;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.EntityDamageSource;
 import net.minecraft.world.entity.Entity;
@@ -17,14 +16,22 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import shadows.apotheosis.Apotheosis;
 import shadows.apotheosis.adventure.affix.Affix;
-import shadows.apotheosis.adventure.affix.AffixHelper;
 import shadows.apotheosis.adventure.affix.AffixType;
+import shadows.apotheosis.adventure.affix.socket.gem.bonus.GemBonus;
 import shadows.apotheosis.adventure.loot.LootCategory;
 import shadows.apotheosis.adventure.loot.LootRarity;
 import shadows.apotheosis.mixin.LivingEntityInvoker;
 import shadows.placebo.util.StepFunction;
 
 public class ExecutingAffix extends Affix {
+
+	//Formatter::off
+	public static final Codec<ExecutingAffix> CODEC = RecordCodecBuilder.create(inst -> inst
+		.group(
+			GemBonus.VALUES_CODEC.fieldOf("values").forGetter(a -> a.values))
+			.apply(inst, ExecutingAffix::new)
+		);
+	//Formatter::on
 
 	protected final Map<LootRarity, StepFunction> values;
 
@@ -64,24 +71,6 @@ public class ExecutingAffix extends Affix {
 				}
 			}
 		}
-	}
-
-	public static Affix read(JsonObject obj) {
-		var values = AffixHelper.readValues(GsonHelper.getAsJsonObject(obj, "values"));
-		return new ExecutingAffix(values);
-	}
-
-	public JsonObject write() {
-		return new JsonObject();
-	}
-
-	public void write(FriendlyByteBuf buf) {
-		buf.writeMap(this.values, (b, key) -> b.writeUtf(key.id()), (b, func) -> func.write(b));
-	}
-
-	public static Affix read(FriendlyByteBuf buf) {
-		Map<LootRarity, StepFunction> values = buf.readMap(b -> LootRarity.byId(b.readUtf()), b -> StepFunction.read(b));
-		return new ExecutingAffix(values);
 	}
 
 }