+++ src/main/java/shadows/apotheosis/adventure/affix/effect/TelepathicAffix.java
@@ -4,6 +4,8 @@ import java.util.Map;
 import java.util.function.Consumer;
 
 import com.google.gson.JsonObject;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.FriendlyByteBuf;
@@ -28,6 +30,14 @@ import shadows.apotheosis.adventure.loot.LootRarity;
  */
 public class TelepathicAffix extends Affix {
 
+	//Formatter::off
+	public static final Codec<TelepathicAffix> CODEC = RecordCodecBuilder.create(inst -> inst
+		.group(
+			LootRarity.CODEC.fieldOf("min_rarity").forGetter(a -> a.minRarity))
+			.apply(inst, TelepathicAffix::new)
+		);
+	//Formatter::on
+	
 	public static Vec3 blockDropTargetPos = null;
 
 	protected LootRarity minRarity;