+++ src/main/java/shadows/apotheosis/ench/table/KeepNBTEnchantingRecipe.java
@@ -1,7 +1,8 @@
 package shadows.apotheosis.ench.table;
 
+import org.apache.commons.lang3.tuple.Pair;
+
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -36,12 +37,8 @@ public class KeepNBTEnchantingRecipe extends EnchantingRecipe {
 		public KeepNBTEnchantingRecipe fromJson(ResourceLocation id, JsonObject obj) {
 			ItemStack output = CraftingHelper.getItemStack(obj.get("result").getAsJsonObject(), true, true);
 			Ingredient input = Ingredient.fromJson(obj.get("input"));
-			Stats stats = GSON.fromJson(obj.get("requirements"), Stats.class);
-			Stats maxStats = obj.has("max_requirements") ? GSON.fromJson(obj.get("max_requirements"), Stats.class) : NO_MAX;
-			if (maxStats.eterna != -1 && stats.eterna > maxStats.eterna) throw new JsonParseException("An enchanting recipe (" + id + ") has invalid min/max eterna bounds (min > max).");
-			if (maxStats.quanta != -1 && stats.quanta > maxStats.quanta) throw new JsonParseException("An enchanting recipe (" + id + ") has invalid min/max quanta bounds (min > max).");
-			if (maxStats.arcana != -1 && stats.arcana > maxStats.arcana) throw new JsonParseException("An enchanting recipe (" + id + ") has invalid min/max arcana bounds (min > max).");
-			return new KeepNBTEnchantingRecipe(id, output, input, stats, maxStats);
+			Pair<Stats, Stats> requirements = readStats(id, obj);
+			return new KeepNBTEnchantingRecipe(id, output, input, requirements.getLeft(), requirements.getRight());
 		}
 
 	}